PyG-Neo4j/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
cycler==0.11.0
decorator==4.4.2
filelock==3.4.2
fonttools==4.29.0
googledrivedownloader==0.4
huggingface-hub==0.4.0
idna==3.3
imageio==2.14.1
imageio-ffmpeg==0.4.5
isodate==0.6.1
Jinja2==3.0.3
joblib==1.1.0
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.5.1
moviepy==1.0.3
neo4j==4.4.1
networkx==2.6.3
nltk==3.6.7
numpy==1.22.1
packaging==21.3
pandas==1.4.0
Pillow==9.0.0
proglog==0.1.9
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
rdflib==6.1.1
regex==2022.1.18
requests==2.27.1
sacremoses==0.0.47
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
sentence-transformers==2.1.0
sentencepiece==0.1.96
six==1.16.0
threadpoolctl==3.0.0
tokenizers==0.11.4
torch==1.10.2
torch-cluster==1.5.9
torch-geometric==2.0.3
torch-scatter==2.0.9
torch-sparse==0.6.12
torch-spline-conv==1.2.1
torchvision==0.11.3
tqdm==4.62.3
transformers==4.16.1
typing-extensions==4.0.1
urllib3==1.26.8
yacs==0.1.8
